+++ superset-frontend/src/SqlLab/components/SouthPane/Results.tsx
@@ -37,7 +37,7 @@ type Props = {
 
 const StyledEmptyStateWrapper = styled.div`
   height: 100%;
-  .ant-empty-image img {
+  .antd5-empty-image img {
     margin-right: 28px;
   }
